Re-introducing the k3OS Operator
Initially focused with providing a kubernetes-native upgrade experience, this change-set enhances the k3os multi-call binary to provide:
- `k3os operator agent` custom resource controller(s)
- `k3os upgrade` perform rootfs and kernel upgrades

New Custom Resources:

- Channel (for managing where to resolve the "latest" version)
- UpgradeSet (for specifying cluster-wide upgrades)
- NodeUpgrade (for specifying node-upgrades)

Design Assumptions:

The resolution of the "latest" version for a `Channel` is very simple in that you specify a URL, such as https://github.com/rancher/k3os/releases/latest, which responds with an HTTP 302 + Location where the base path, aka the last path element, of the Location is a release tag. This version is placed into `ChannelStatus.Version` which will trigger the creation of an `UpgradeSet`.

`UpgradeSets` are automatically created for `Channels` but a `Channel` is not a necessary predicate for an `UpgradeSet`.  They represent an upgrade policy for k3OS nodes in the cluster by specifying and tracking the number of concurrent upgrades that can run.

`NodeUpgrades` reify an `UpgradeSet` for a `Node` specified by name. They are triggered on the creation/update of an `UpgradeSet` but may be created manually to incur a one-time upgrade. They trigger the creation and track the progress of a node-bound batch `Job` to perform the upgrade on the node.