Updated parted args to correctly grow the partition to 100%
The parted command wasn't resizing the root partition to use the available disk space.  The issue in parted is also documented in the bug report https://bugs.launchpad.net/ubuntu/+source/parted/+bug/1270203. 

In this PR the invocation of the `parted` command has been updated so that the partition is properly resized to used the available disk space.